footballApp/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-enumchoicefield==2.0.0
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
